requirements.txt
tensorflow
transformers
flask-cors
numpy
sklearn
pytorch
vncorenlp
gunicorn
click
Flask
itsdangerous
Jinja2
MarkupSafe
Werkzeug